Fix horizontal scrolling in plug drawer.
