+++ src/app/compare/Compare.m.scss.d.ts
@@ -4,7 +4,6 @@ interface CssExports {
   'bucket': string;
   'fixedLeft': string;
   'highlightBar': string;
-  'items': string;
   'options': string;
   'organizerLink': string;
   'sorted': string;
