+++ src/app/compare/Compare.tsx
@@ -15,16 +15,15 @@ import { useSetting } from 'app/settings/hooks';
 import { AppIcon, faAngleLeft, faAngleRight, faList } from 'app/shell/icons';
 import { acquisitionRecencyComparator } from 'app/shell/item-comparators';
 import { useThunkDispatch } from 'app/store/thunk-dispatch';
-import { isEventFromFirefoxScrollbar } from 'app/utils/browsers';
 import { emptyArray } from 'app/utils/empty';
 import { DestinyDisplayPropertiesDefinition } from 'bungie-api-ts/destiny2';
 import clsx from 'clsx';
 import { StatHashes } from 'data/d2/generated-enums';
 import _ from 'lodash';
-import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import { useSelector } from 'react-redux';
 import { Link } from 'react-router-dom';
-import Sheet from '../dim-ui/Sheet';
+import Sheet, { HorizontalScrollContainer } from '../dim-ui/Sheet';
 import { DimItem, DimSocket } from '../inventory/item-types';
 import { chainComparator, compareBy, reverseComparator } from '../utils/comparators';
 import styles from './Compare.m.scss';
@@ -260,9 +259,6 @@ export default function Compare({ session }: { session: CompareSession }) {
   );
 }
 
-// After this many pixels of dragging in either direction, we consider ourselves to be part of a scrolling gesture.
-const HORIZ_SCROLL_DRAG_THRESHOLD = 20;
-
 function CompareItems({
   items,
   doCompareBaseStats,
@@ -280,66 +276,8 @@ function CompareItems({
   setHighlight: React.Dispatch<React.SetStateAction<string | number | undefined>>;
   onPlugClicked: (value: { item: DimItem; socket: DimSocket; plugHash: number }) => void;
 }) {
-  // This uses pointer events to directly set the scroll position based on
-  // dragging the items. This works around an iOS bug around nested draggables,
-  // but also is kinda nice on desktop. I wasn't able to get it to do an
-  // inertial animation after releasing.
-
-  const ref = useRef<HTMLDivElement>(null);
-  const dragStateRef = useRef<{
-    scrollPosition: number;
-    pointerDownPosition: number;
-    scrolling: boolean;
-  }>();
-  const handlePointerDown = useCallback((e: React.PointerEvent) => {
-    if (isEventFromFirefoxScrollbar(e)) {
-      return;
-    }
-
-    // Don't do any of this if the view isn't scrollable in the first place
-    if (ref.current!.scrollWidth <= ref.current!.clientWidth) {
-      return;
-    }
-
-    dragStateRef.current = {
-      pointerDownPosition: e.clientX,
-      scrollPosition: ref.current!.scrollLeft,
-      scrolling: false,
-    };
-  }, []);
-  const handlePointerUp = useCallback((e: React.PointerEvent) => {
-    dragStateRef.current = undefined;
-    ref.current!.releasePointerCapture(e.pointerId);
-  }, []);
-  const handlePointerMove = useCallback((e: React.PointerEvent) => {
-    if (dragStateRef.current !== undefined) {
-      const { scrollPosition, pointerDownPosition } = dragStateRef.current;
-      // Once we've moved HORIZ_SCROLL_DRAG_THRESHOLD in either direction,
-      // constrain to horizontal scrolling only
-      dragStateRef.current.scrolling ||=
-        Math.abs(e.clientX - pointerDownPosition) > HORIZ_SCROLL_DRAG_THRESHOLD;
-      if (dragStateRef.current.scrolling) {
-        // Only set the pointer capture once we've moved enough. This allows you
-        // to still keep scrolling even if the pointer leaves the scrollable
-        // area (which feels nice) but buttons still work. If we always capture
-        // in handlePointerDown, buttons won't work because all events get
-        // retargeted to the scroll area.
-        ref.current!.setPointerCapture(e.pointerId);
-        e.stopPropagation();
-      }
-      ref.current!.scrollLeft = scrollPosition - (e.clientX - pointerDownPosition);
-    }
-  }, []);
-
   return (
-    <div
-      ref={ref}
-      className={styles.items}
-      onPointerDown={handlePointerDown}
-      onPointerMove={handlePointerMove}
-      onPointerUp={handlePointerUp}
-      onPointerCancel={handlePointerUp}
-    >
+    <HorizontalScrollContainer>
       {items.map((item) => (
         <CompareItem
           item={item}
@@ -353,7 +291,7 @@ function CompareItems({
           isInitialItem={initialItemId === item.id}
         />
       ))}
-    </div>
+    </HorizontalScrollContainer>
   );
 }
