+++ src/app/dim-ui/Sheet.m.scss
@@ -1,20 +1,33 @@
+@import '../variables';
+
 .sheetDisabled {
   transform-origin: center bottom;
   // TODO: would be better to do with Framer Motion, once we switch over to it
   transform: scale(0.98) !important;
   transition: transform 300ms linear;
 }
+
 .disabledScreen {
   background-color: black;
   position: absolute;
-  left: 0;
-  right: 0;
-  top: 0;
-  bottom: 0;
+  inset: 0 0 0 0;
   display: none;
   transition: opacity 300ms linear;
   .sheetDisabled & {
     display: block;
     opacity: 0.6;
   }
+}
+
+.horizontalScrollContainer {
+  flex: 1;
+  white-space: nowrap;
+  overflow-x: scroll;
+  overflow-y: hidden;
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-start;
+  -webkit-overflow-scrolling: touch;
+  contain: content;
+  touch-action: none;
 }
