+++ src/app/dim-ui/Sheet.m.scss.d.ts
@@ -2,6 +2,7 @@
 // Please do not change this file!
 interface CssExports {
   'disabledScreen': string;
+  'horizontalScrollContainer': string;
   'sheetDisabled': string;
 }
 export const cssExports: CssExports;
