+++ src/app/loadout/plug-drawer/Footer.m.scss.d.ts
@@ -2,7 +2,6 @@
 // Please do not change this file!
 interface CssExports {
   'footer': string;
-  'selectedPlugs': string;
   'submitButton': string;
 }
 export const cssExports: CssExports;
